requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-firestore
google-cloud-aiplatform
python-multipart
python-dotenv
requests